chore: refactor esr handling
